Refactor control flow and imports across app
