+++ src/components/settings/sections/AppearanceSection.tsx
@@ -10,8 +10,8 @@ import { useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
 import { Palette } from 'lucide-react'
 
-import { SelectSettingsSection } from '../primitives'
-import type { SelectOption } from '../primitives'
+import { SelectSettingsSection } from '../primitives/SelectSettingsSection'
+import type { SelectOption } from '../primitives/SettingsSelect'
 import type { Theme } from '@/stores/app-store'
 import { useAppStore } from '@/stores/app-store'
